cloud_python/requirements.txt
google-cloud-vision==2.0.0
google-cloud-storage==1.35.0
Wand==0.6.5